requirements.txt
openai-whisper==20231117
ollama==0.1.7
moviepy==1.0.3
torch>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0
rich>=13.0.0
playsound>=1.3.0

# Speaker diarization dependencies
pyannote.audio @ git+https://github.com/pyannote/pyannote-audio.git
pyannote.core
pyannote.database
pyannote.metrics
librosa>=0.9.0
soundfile>=0.10.0